Initial refactor of streaming code
It still doesn't do everything I want or look quite right but the
functionality is mostly there.

- Can stream zipfile reads - proof in the progressbar
- All of the streaming stuff is more consistent
- Also more concise
- Ok, so technically there's more code but it's (slightly) easier to
understand
- Added a couple of tests (need more) (also, sloooooow)
- Cleanup.

I'm really aiming for this to be simple:
  reader = Input.new(source)
  reader.add_transform(foo)
  reader.add_transform(bar)
  reader.go
where the Input is just figured out, the transforms including writing
and hashing and progress bars, and something like "go" or "do_it" just
makes everything happen. Possibly a rewind method, too. But I couldn't
find that in a consistent way for all the components I wanted.